[MOVE] : util 파일 common 모듈로 이전 #415
