+++ app/src/main/java/org/sopt/official/feature/auth/AuthActivity.kt
@@ -42,25 +42,25 @@ import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
+import org.sopt.officiail.domain.mypage.model.UserStatus
 import org.sopt.official.BuildConfig
 import org.sopt.official.R
 import org.sopt.official.auth.PlaygroundAuth
 import org.sopt.official.auth.data.PlaygroundAuthDatasource
-import org.sopt.official.network.model.response.OAuthToken
 import org.sopt.official.common.di.Auth
+import org.sopt.official.common.util.dp
+import org.sopt.official.common.util.setOnAnimationEndListener
+import org.sopt.official.common.util.setOnSingleClickListener
+import org.sopt.official.common.util.viewBinding
 import org.sopt.official.config.messaging.SoptFirebaseMessagingService.Companion.REMOTE_MESSAGE_EVENT_LINK
 import org.sopt.official.config.messaging.SoptFirebaseMessagingService.Companion.REMOTE_MESSAGE_EVENT_TYPE
-import org.sopt.official.data.mypage.remote.model.request.AuthRequest
 import org.sopt.official.data.mypage.remote.api.AuthService
+import org.sopt.official.data.mypage.remote.model.request.AuthRequest
 import org.sopt.official.databinding.ActivityAuthBinding
-import org.sopt.officiail.domain.mypage.model.UserStatus
 import org.sopt.official.feature.home.HomeActivity
+import org.sopt.official.feature.toEntity
+import org.sopt.official.network.model.response.OAuthToken
 import org.sopt.official.network.persistence.SoptDataStore
-import org.sopt.official.feature.mypage.util.dp
-import org.sopt.official.util.setOnAnimationEndListener
-import org.sopt.official.util.setOnSingleClickListener
-import org.sopt.official.util.toEntity
-import org.sopt.official.feature.mypage.util.viewBinding
 import javax.inject.Inject
 
 @AndroidEntryPoint
