+++ app/src/main/java/org/sopt/official/feature/home/adapter/SmallBlockAdapter.kt
@@ -31,12 +31,12 @@ import android.view.ViewGroup
 import androidx.core.view.isVisible
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
+import org.sopt.official.common.util.drawableOf
+import org.sopt.official.common.util.setOnSingleClickListener
+import org.sopt.official.common.util.stringOf
 import org.sopt.official.common.view.ItemDiffCallback
 import org.sopt.official.databinding.ItemMainSmallBlockListBinding
 import org.sopt.official.feature.home.model.HomeMenuType
-import org.sopt.official.feature.mypage.util.drawableOf
-import org.sopt.official.util.setOnSingleClickListener
-import org.sopt.official.feature.mypage.util.stringOf
 
 class SmallBlockAdapter : ListAdapter<HomeMenuType, SmallBlockAdapter.ViewHolder>(
     ItemDiffCallback(
