--- app/src/main/java/org/sopt/official/util/GenerationExt.kt
@@ -1,47 +1,0 @@
-/*
- * MIT License
- * Copyright 2023 SOPT - Shout Our Passion Together
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package org.sopt.official.util
-
-import kotlinx.datetime.DateTimeUnit
-import kotlinx.datetime.LocalDate
-import kotlinx.datetime.periodUntil
-import kotlinx.datetime.plus
-import kotlinx.datetime.toLocalDate
-
-private const val BASE_DATE = "2007-03-01"
-
-fun calculateGenerationStartDate(
-    generation: Int,
-    period: Int = 6,
-    from: LocalDate = BASE_DATE.toLocalDate()
-): LocalDate {
-    val monthsToTake = generation * period
-    return from.plus(monthsToTake, DateTimeUnit.MONTH)
-}
-
-fun calculateDurationOfGeneration(start: LocalDate, end: LocalDate): Int {
-    val diff = start.periodUntil(end)
-    return diff.months + diff.years * 12
-}
