--- app/src/main/java/org/sopt/official/util/Intent.kt
@@ -1,59 +1,0 @@
-/*
- * MIT License
- * Copyright 2023 SOPT - Shout Our Passion Together
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package org.sopt.official.util
-
-import android.app.Activity
-import android.content.Intent
-import android.os.Build
-import android.os.Parcelable
-import java.io.Serializable
-import kotlin.properties.ReadOnlyProperty
-
-fun stringExtra(defaultValue: String? = null) =
-    ReadOnlyProperty<Activity, String?> { thisRef, property ->
-        if (defaultValue == null) {
-            thisRef.intent.extras?.getString(property.name)
-        } else {
-            thisRef.intent.extras?.getString(property.name, defaultValue)
-        }
-    }
-
-inline fun <reified T : Serializable?> Intent.serializableExtra(key: String): T? {
-    return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-        getSerializableExtra(key, T::class.java)
-    } else {
-        getSerializableExtra(key) as? T
-    }
-}
-
-inline fun <reified S : Serializable> serializableExtra(defaultValue: S? = null) =
-    ReadOnlyProperty<Activity, S?> { thisRef, property ->
-        thisRef.intent.serializableExtra(property.name) ?: defaultValue
-    }
-
-inline fun <reified T : Parcelable> parcelableExtra(defaultValue: T? = null) =
-    ReadOnlyProperty<Activity, T?> { thisRef, property ->
-        thisRef.intent.getParcelableExtra(property.name) ?: defaultValue
-    }
