+++ core/common/src/main/java/org/sopt/official/common/util/Binding.kt
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.sopt.official.util
+package org.sopt.official.common.util
 
 import android.view.LayoutInflater
 import android.view.View
