+++ core/common/src/main/java/org/sopt/official/common/util/rx/SingleExt.kt
@@ -22,18 +22,17 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.sopt.official.util.ui
+package org.sopt.official.common.util.rx
 
-import io.reactivex.rxjava3.core.BackpressureStrategy
-import io.reactivex.rxjava3.core.Flowable
-import io.reactivex.rxjava3.core.Observable
-import org.sopt.official.feature.mypage.util.ui.UxConstant
-import java.util.concurrent.TimeUnit
+import io.reactivex.rxjava3.core.Single
+import io.reactivex.rxjava3.disposables.CompositeDisposable
+import io.reactivex.rxjava3.schedulers.Schedulers
+import timber.log.Timber
 
-object UxConstant {
-    const val UI_REACTION_TIME = 300L
-}
+fun <T : Any> Single<T>.subscribeOnIo() = this.subscribeOn(Schedulers.io())
 
-fun <T : Any> Observable<T>.throttleUi(interval: Long = UxConstant.UI_REACTION_TIME): Flowable<T> = this
-    .toFlowable(BackpressureStrategy.DROP)
-    .throttleFirst(interval, TimeUnit.MILLISECONDS)
+fun <T : Any> Single<T>.subscribeBy(
+    compositeDisposable: CompositeDisposable,
+    onError: (Throwable) -> Unit = { Timber.e(it.message) },
+    onSuccess: (T) -> Unit
+) = compositeDisposable.add(subscribe(onSuccess, onError))
