+++ feature/mypage/src/main/java/org/sopt/official/feature/mypage/soptamp/nickName/ChangeNickNameActivity.kt
@@ -34,15 +34,15 @@ import com.jakewharton.rxbinding4.widget.textChanges
 import dagger.hilt.android.AndroidEntryPoint
 import io.reactivex.rxjava3.core.BackpressureStrategy
 import io.reactivex.rxjava3.disposables.CompositeDisposable
+import org.sopt.official.common.util.drawableOf
+import org.sopt.official.common.util.rx.observeOnMain
+import org.sopt.official.common.util.rx.subscribeBy
+import org.sopt.official.common.util.rx.subscribeOnIo
+import org.sopt.official.common.util.ui.setVisible
+import org.sopt.official.common.util.ui.throttleUi
+import org.sopt.official.common.util.viewBinding
 import org.sopt.official.feature.mypage.R
 import org.sopt.official.feature.mypage.databinding.ActivityChangeNickNameBinding
-import org.sopt.official.feature.mypage.util.drawableOf
-import org.sopt.official.feature.mypage.util.rx.observeOnMain
-import org.sopt.official.feature.mypage.util.rx.subscribeBy
-import org.sopt.official.feature.mypage.util.rx.subscribeOnIo
-import org.sopt.official.feature.mypage.util.ui.setVisible
-import org.sopt.official.feature.mypage.util.ui.throttleUi
-import org.sopt.official.feature.mypage.util.viewBinding
 
 @AndroidEntryPoint
 class ChangeNickNameActivity : AppCompatActivity() {
