--- feature/mypage/src/main/java/org/sopt/official/feature/mypage/util/Binding.kt
@@ -1,84 +1,0 @@
-/*
- * MIT License
- * Copyright 2023 SOPT - Shout Our Passion Together
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package org.sopt.official.feature.mypage.util
-
-import android.view.LayoutInflater
-import android.view.View
-import androidx.appcompat.app.AppCompatActivity
-import androidx.fragment.app.Fragment
-import androidx.lifecycle.DefaultLifecycleObserver
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.LifecycleOwner
-import androidx.viewbinding.ViewBinding
-import kotlin.properties.ReadOnlyProperty
-import kotlin.reflect.KProperty
-
-inline fun <T : ViewBinding> AppCompatActivity.viewBinding(
-    crossinline inflater: (LayoutInflater) -> T
-) = lazy(LazyThreadSafetyMode.NONE) {
-    inflater.invoke(layoutInflater)
-}
-
-class FragmentViewBindingDelegate<F : Fragment, T : ViewBinding>(
-    val fragment: F,
-    val viewBindingFactory: (View) -> T
-) : ReadOnlyProperty<F, T> {
-    private var binding: T? = null
-
-    init {
-        fragment.lifecycle.addObserver(
-            object : DefaultLifecycleObserver {
-                override fun onCreate(owner: LifecycleOwner) {
-                    fragment.viewLifecycleOwnerLiveData.observe(fragment) { viewLifecycleOwner ->
-                        viewLifecycleOwner.lifecycle.addObserver(
-                            object : DefaultLifecycleObserver {
-                                override fun onDestroy(owner: LifecycleOwner) {
-                                    binding = null
-                                }
-                            }
-                        )
-                    }
-                }
-            }
-        )
-    }
-
-    override fun getValue(thisRef: F, property: KProperty<*>): T {
-        val binding = binding
-        if (binding != null) {
-            return binding
-        }
-
-        val lifecycle = fragment.viewLifecycleOwner.lifecycle
-        if (!lifecycle.currentState.isAtLeast(Lifecycle.State.INITIALIZED)) {
-            throw IllegalStateException("Should not attempt to get bindings when Fragment views are destroyed.")
-        }
-
-        return viewBindingFactory(thisRef.requireView()).also { this.binding = it }
-    }
-}
-
-fun <T : ViewBinding> Fragment.viewBinding(viewBinder: (View) -> T) =
-    FragmentViewBindingDelegate(this, viewBinder)
